requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
typer>=0.12,<1.0
rich>=13.9,<14.0
requests>=2.32,<3.0
psutil>=6.1,<7.0
email-validator>=2.2,<3.0
google-genai>=1.15,<2.0
openai>=1.40,<2.0
Pillow>=12.2,<13.0
winrt-runtime>=3.2,<4.0
winrt-Windows.Foundation>=3.2,<4.0
winrt-Windows.Foundation.Collections>=3.2,<4.0
winrt-Windows.Globalization>=3.2,<4.0
winrt-Windows.Graphics.Imaging>=3.2,<4.0
winrt-Windows.Media.Ocr>=3.2,<4.0
winrt-Windows.Storage>=3.2,<4.0
winrt-Windows.Storage.Streams>=3.2,<4.0
